006_bedtime_story_teller/requirements.txt
langchain==0.0.247
python-dotenv==1.0.0
streamlit==1.25.0
openai==0.27.8